# sprint-5/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kognos API",
        "description": "Documentación de la API REST para la plataforma Kognos.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/courses/{course}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get full details of a course by ID",
                "operationId": "fee12cdd5ac7c568500478014248cd1d",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "ID of the course to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Test Course"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "This is a test course"
                                                },
                                                "content": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Capítulo 1"
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "example": "Descripción del capítulo 1"
                                                            },
                                                            "videos": {
                                                                "type": "array",
                                                                "items": {
                                                                    "properties": {
                                                                        "title": {
                                                                            "type": "string",
                                                                            "example": "Test Video 1"
                                                                        },
                                                                        "description": {
                                                                            "type": "string",
                                                                            "example": "This is a test video 1"
                                                                        },
                                                                        "url": {
                                                                            "type": "string",
                                                                            "format": "url",
                                                                            "example": "https://www.youtube.com/watch?v=123456"
                                                                        }
                                                                    },
                                                                    "type": "object"
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Course not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Admin - Courses"
                ],
                "summary": "Update an existing course (Admin only)",
                "operationId": "453c1cb3d7a8cadde3674a242bceb96e",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "ID of the course to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Course"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated course description"
                                    },
                                    "content": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Updated Chapter 1"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Updated chapter description"
                                                },
                                                "videos": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Video 1"
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "example": "Updated video description"
                                                            },
                                                            "url": {
                                                                "type": "string",
                                                                "format": "url",
                                                                "example": "https://www.youtube.com/watch?v=video1"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Updated Course"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Updated course description"
                                        },
                                        "content": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized – missing or invalid token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden – user does not have admin role or permission",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User does not have the right roles."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Course] 12"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "content.0.videos.0.title": [
                                                    "The content.0.videos.0.title must be a string."
                                                ],
                                                "content.0.videos.0.description": [
                                                    "The content.0.videos.0.description must be a string."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Courses"
                ],
                "summary": "Delete a course (Admin only)",
                "operationId": "7b645feb29b6dc51ef393e0ef5ddecb1",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "ID of the course to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Course deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized – missing or invalid token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden – user does not have admin role or permission",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User does not have the right roles."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Course] 99"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
...
}